+++ src/wai/common/json/configuration/property/__init__.py
@@ -6,6 +6,7 @@ from ._ArrayProperty import ArrayProperty
 from ._BoolProperty import BoolProperty
 from ._ConstantProperty import ConstantProperty
 from ._EnumProperty import EnumProperty
+from ._MapProperty import MapProperty
 from ._NumberProperty import NumberProperty
 from ._OneOfProperty import OneOfProperty
 from ._Property import Property